requirements.txt
Django==3.1
ispyb
djangorestframework
psycopg2-binary
django-dbconn-retry
sshtunnel
PyMySQL
django_cas_ng
django-guardian